MILESTONE 4/requirements.txt
# E2E Testing Agent - Requirements (Milestone 4)
# Python 3.8+

# Core Framework
flask==3.0.0
python-dotenv==1.0.0

# Browser Automation
playwright==1.40.0

# LangChain & AI
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0
langchain-community==0.0.13

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Optional: For advanced features
# selenium==4.15.2
# beautifulsoup4==4.12.2
# pillow==10.1.0